requirements.txt
opencv-python
SpeechRecognition
pyttsx3
pyaudio
tk 